requirements.txt
# Web Framework
flask==3.0.2

# Database
sqlalchemy==2.0.27
pymysql==1.1.0
cryptography==42.0.2

# AI/ML
langchain-core==0.1.27
langchain-community==0.0.24
langchain-openai==0.0.5
openai==1.12.0

# Enhanced Matching System
sentence-transformers==2.2.2
scikit-learn==1.4.0
numpy==1.26.3
redis==5.0.1
spacy==3.7.2

# Utilities
python-dotenv==1.0.1
pandas==2.2.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
pytest==8.0.1 
bcrypt==4.1.2
PyJWT==2.8.0 